Move occurrence test queries and mutations to new files
Refactored the occurrence API tests so that queries and mutations are in their own files
